Fix #133: Promote measurement to a top-level feature
Move measurement domain and data out of settings into feature:measurement,
wire search, recipe details, cooking, and settings UI to the new modules,
use Local/Remote data sources with Metro ContributesBinding, and split
aggregated use case files into one class per file.
