+++ app/build.gradle.kts
@@ -90,6 +90,7 @@ dependencies {
 	implementation(project(":feature:newrecipe:data"))
 	implementation(project(":feature:recipedetails:data"))
 	implementation(project(":feature:search:data"))
+	implementation(project(":feature:measurement:data"))
 	implementation(project(":feature:settings:data"))
 	implementation(project(":shared:data"))
 	implementation(project(":shared:ui"))
