+++ feature/measurement/data/src/androidMain/kotlin/app/purecipes/feature/measurement/data/datasource/PlatformRegionCode.android.kt
@@ -1,0 +1,7 @@
+package app.purecipes.feature.measurement.data.datasource
+
+import java.util.Locale
+
+internal actual fun detectRegionCode(): String? {
+	return Locale.getDefault().country.takeIf { it.isNotBlank() }
+}
