+++ feature/measurement/data/src/commonMain/kotlin/app/purecipes/feature/measurement/data/datasource/PlatformRegionCode.kt
@@ -1,0 +1,3 @@
+package app.purecipes.feature.measurement.data.datasource
+
+internal expect fun detectRegionCode(): String?
