+++ feature/measurement/data/src/jvmMain/kotlin/app/purecipes/feature/measurement/data/datasource/PlatformRegionCode.jvm.kt
@@ -1,0 +1,7 @@
+package app.purecipes.feature.measurement.data.datasource
+
+import java.util.Locale
+
+internal actual fun detectRegionCode(): String? {
+	return Locale.getDefault().country.takeIf { it.isNotBlank() }
+}
