+++ feature/measurement/data/src/wasmJsMain/kotlin/app/purecipes/feature/measurement/data/datasource/PlatformRegionCode.wasmJs.kt
@@ -1,0 +1,5 @@
+package app.purecipes.feature.measurement.data.datasource
+
+internal actual fun detectRegionCode(): String? {
+	return null
+}
