+++ feature/measurement/domain/build.gradle.kts
@@ -1,0 +1,18 @@
+plugins {
+	id("convention.domain")
+	id("convention.common-test")
+}
+
+kotlin {
+	android {
+		namespace = "app.purecipes.feature.measurement.domain"
+	}
+
+	sourceSets {
+		commonMain {
+			dependencies {
+				api(project(":shared:domain"))
+			}
+		}
+	}
+}
