+++ feature/measurement/domain/src/commonMain/kotlin/app/purecipes/feature/measurement/domain/model/ProcessedRecipeDetails.kt
