+++ feature/measurement/domain/src/commonMain/kotlin/app/purecipes/feature/measurement/domain/usecase/SaveMeasurementPreferencesUseCase.kt
@@ -1,0 +1,15 @@
+package app.purecipes.feature.measurement.domain.usecase
+
+import app.purecipes.feature.measurement.domain.repository.MeasurementPreferencesRepository
+import app.purecipes.shared.domain.model.MeasurementPreferences
+import dev.zacsweers.metro.Inject
+
+@Inject
+class SaveMeasurementPreferencesUseCase(
+	private val repository: MeasurementPreferencesRepository,
+) {
+
+	suspend operator fun invoke(preferences: MeasurementPreferences) {
+		repository.saveMeasurementPreferences(preferences)
+	}
+}
