+++ feature/recipedetails/ui/build.gradle.kts
@@ -21,7 +21,7 @@ kotlin {
 				api(project(":feature:analytics:domain"))
 				api(project(":feature:favorites:domain"))
 				api(project(":feature:recipedetails:domain"))
-				api(project(":feature:settings:domain"))
+				api(project(":feature:measurement:domain"))
 				api(project(":feature:sharing:domain"))
 				implementation(project(":feature:sharing:ui"))
 				api(project(":shared:domain"))
