+++ feature/search/ui/build.gradle.kts
@@ -19,7 +19,7 @@ kotlin {
 			dependencies {
 				api(project(":feature:analytics:domain"))
 				api(project(":feature:search:domain"))
-				api(project(":feature:settings:domain"))
+				api(project(":feature:measurement:domain"))
 				api(project(":shared:domain"))
 				implementation(libs.coil.compose)
 				implementation(libs.coil.networkKtor3)
