--- feature/settings/data/src/commonMain/kotlin/app/purecipes/feature/settings/data/datasource/PlatformRegionCode.kt
@@ -1,3 +1,0 @@
-package app.purecipes.feature.settings.data.datasource
-
-internal expect fun detectRegionCode(): String?
