--- feature/settings/data/src/commonMain/kotlin/app/purecipes/feature/settings/data/repository/MeasurementPreferencesAccessor.kt
@@ -1,92 +1,0 @@
-package app.purecipes.feature.settings.data.repository
-
-import app.purecipes.feature.measurement.domain.repository.MeasurementPreferencesRepository
-import app.purecipes.feature.settings.data.datasource.MeasurementPreferencesDataSource
-import app.purecipes.feature.settings.data.datasource.MeasurementPreferencesRemoteDataSource
-import app.purecipes.shared.data.session.SessionTokenStore
-import app.purecipes.shared.domain.model.MeasurementPreferences
-import kotlinx.coroutines.NonCancellable
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.emitAll
-import kotlinx.coroutines.flow.flow
-import kotlinx.coroutines.withContext
-
-class MeasurementPreferencesAccessor(
-	private val localDataSource: MeasurementPreferencesDataSource,
-	private val remoteDataSource: MeasurementPreferencesRemoteDataSource,
-	private val sessionTokenStore: SessionTokenStore,
-) : MeasurementPreferencesRepository {
-
-	override fun observeMeasurementPreferences(): Flow<MeasurementPreferences> {
-		return flow {
-			syncRemoteToLocalIfAuthenticated()
-			emitAll(localDataSource.observeMeasurementPreferences())
-		}
-	}
-
-	override suspend fun getMeasurementPreferences(): MeasurementPreferences {
-		val localPreferences = localDataSource.getMeasurementPreferences()
-		if (!isAuthenticated()) {
-			return localPreferences
-		}
-
-		val result = remoteDataSource.getMeasurementPreferences()
-		if (result.isOk) {
-			return result.component1()
-				?.also(localDataSource::saveMeasurementPreferences)
-				?: localPreferences
-		}
-
-		syncLocalToRemote(localPreferences)
-		return localPreferences
-	}
-
-	override suspend fun saveMeasurementPreferences(preferences: MeasurementPreferences) {
-		localDataSource.saveMeasurementPreferences(preferences)
-		withContext(NonCancellable) {
-			syncLocalToRemote(preferences)
-		}
-	}
-
-	override suspend fun resetMeasurementPreferences() {
-		localDataSource.resetMeasurementPreferences()
-		withContext(NonCancellable) {
-			syncLocalToRemote(localDataSource.getMeasurementPreferences())
-		}
-	}
-
-	override suspend fun markMismatchNotificationSeen(recipeId: Int) {
-		localDataSource.markMismatchNotificationSeen(recipeId)
-		withContext(NonCancellable) {
-			syncLocalToRemote(localDataSource.getMeasurementPreferences())
-		}
-	}
-
-	private suspend fun syncRemoteToLocalIfAuthenticated() {
-		if (!isAuthenticated()) {
-			return
-		}
-
-		val result = remoteDataSource.getMeasurementPreferences()
-		if (result.isOk) {
-			result.component1()?.let(localDataSource::saveMeasurementPreferences)
-		} else {
-			syncLocalToRemote(localDataSource.getMeasurementPreferences())
-		}
-	}
-
-	private suspend fun syncLocalToRemote(preferences: MeasurementPreferences) {
-		if (!isAuthenticated()) {
-			return
-		}
-
-		val result = remoteDataSource.saveMeasurementPreferences(preferences)
-		if (result.isOk) {
-			result.component1()?.let(localDataSource::saveMeasurementPreferences)
-		}
-	}
-
-	private fun isAuthenticated(): Boolean {
-		return !sessionTokenStore.currentAccessToken().isNullOrBlank()
-	}
-}
