--- feature/settings/data/src/iosMain/kotlin/app/purecipes/feature/settings/data/datasource/PlatformRegionCode.ios.kt
@@ -1,5 +1,0 @@
-package app.purecipes.feature.settings.data.datasource
-
-internal actual fun detectRegionCode(): String? {
-	return null
-}
