--- feature/settings/data/src/jvmMain/kotlin/app/purecipes/feature/settings/data/datasource/PlatformRegionCode.jvm.kt
@@ -1,7 +1,0 @@
-package app.purecipes.feature.settings.data.datasource
-
-import java.util.Locale
-
-internal actual fun detectRegionCode(): String? {
-	return Locale.getDefault().country.takeIf { it.isNotBlank() }
-}
