--- feature/settings/data/src/wasmJsMain/kotlin/app/purecipes/feature/settings/data/datasource/PlatformRegionCode.wasmJs.kt
@@ -1,5 +1,0 @@
-package app.purecipes.feature.settings.data.datasource
-
-internal actual fun detectRegionCode(): String? {
-	return null
-}
