+++ settings.gradle.kts
@@ -40,6 +40,8 @@ include(":feature:auth:ui")
 include(":feature:analytics:data")
 include(":feature:analytics:domain")
 include(":feature:analytics:ui")
+include(":feature:measurement:data")
+include(":feature:measurement:domain")
 include(":feature:main")
 include(":feature:newrecipe:data")
 include(":feature:newrecipe:domain")
